+++ app.js
@@ -8,7 +8,8 @@ var bodyParser = require('body-parser');
 var db = require('./model/db');
 
 var routes = require('./routes/index');
-var users = require('./routes/users');
+var profile = require('./routes/profile');
+var login = require('./routes/login');
 //var contacts = require('./routes/contacts');
 
 var app = express();
@@ -26,7 +27,8 @@ app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
 app.use('/', routes);
-
+app.use('/profiles', profile);
+app.use('/login', login);
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
     var err = new Error('Not Found');
