--- model/conversation.js
@@ -1,8 +1,0 @@
-var mongoose = require('mongoose');
-
-var conversationSchema = new mongoose.Schema({
-    participants:[Profile],
-    messages:[Message]
-});
-
-mongoose.model('Conversation', conversationSchema);
