+++ model/db.js
@@ -66,4 +66,8 @@ process.on('SIGTERM', function() {
 
 // BRING IN YOUR SCHEMAS & MODELS
 // Use plural form of name
-//require('./contacts');
+require('./profiles');
+require('./messages');
+require('./modifications');
+require('./recipes');
+require('./conversations');
