+++ model/modifications.js
@@ -1,7 +1,7 @@
 var mongoose = require('mongoose');
 
 var recipeModificationSchema = new mongoose.Schema({
-    author: Profile,
+    author: {type: mongoose.Schema.Types.ObjectId, ref: 'Profile'},
     ingredientChanges: [{index: Number,name:String, amount:String}],
     stepChanges:[{index: Number, stepDefinition:String}]
 });
