+++ model/profiles.js
@@ -5,8 +5,8 @@ var profileSchema  = new mongoose.Schema({
     lastName: String,
     email: String,
     passwordHash: String,
-    recipes: [ObjectId], //We will perform lookups for these
-    friends: [ObjectId],
+    recipes: [{type: mongoose.Schema.Types.ObjectId, ref: 'Recipe'}], //We will perform lookups for these
+    friends: [{type: mongoose.Schema.Types.ObjectId, ref: 'Profile'}],
     profilePicture:[{data:Buffer, name:String}]
 });
