+++ model/recipes.js
@@ -2,7 +2,7 @@ var mongoose = require('mongoose');
 
 var recipeSchema = new mongoose.Schema({
     name: String,
-    author: Profile,
+    author: {type: mongoose.Schema.Types.ObjectId, ref: 'Profile'},
     description: String,
     ingredients: [{name:String, amount:String}],
     steps: [{stepDefinition: String}],
