+++ npm-debug.log
@@ -1,0 +1,47 @@
+0 info it worked if it ends with ok
+1 verbose cli [ 'C:\\Program Files\\nodejs\\node.exe',
+1 verbose cli   'C:\\Program Files\\nodejs\\node_modules\\npm\\bin\\npm-cli.js',
+1 verbose cli   'start' ]
+2 info using npm@3.10.8
+3 info using node@v6.9.0
+4 verbose run-script [ 'prestart', 'start', 'poststart' ]
+5 info lifecycle recipe-api@1.0.0~prestart: recipe-api@1.0.0
+6 silly lifecycle recipe-api@1.0.0~prestart: no script for prestart, continuing
+7 info lifecycle recipe-api@1.0.0~start: recipe-api@1.0.0
+8 verbose lifecycle recipe-api@1.0.0~start: unsafe-perm in lifecycle true
+9 verbose lifecycle recipe-api@1.0.0~start: PATH: C:\Program Files\nodejs\node_modules\npm\bin\node-gyp-bin;C:\Users\millerlj\Documents\GitHub\CSSE280-Recipe-Api\node_modules\.bin;C:\Python34\;C:\Python34\Scripts;C:\Program Files (x86)\NVIDIA Corporation\PhysX\Common;C:\ProgramData\Oracle\Java\javapath;C:\windows\system32;C:\windows;C:\windows\System32\Wbem;C:\windows\System32\WindowsPowerShell\v1.0\;C:\windows\System32\WindowsPowerShell\v1.0\;C:\Program Files\VanDyke Software\Clients\;C:\Program Files\MATLAB\R2015a\runtime\win64;C:\Program Files\MATLAB\R2015a\bin;C:\Program Files\TortoiseSVN\bin;c:\mingw\bin;C:\Program Files\Lenovo\Touch Fingerprint Software\;C:\Program Files (x86)\Skype\Phone\;C:\Program Files\Git\cmd;C:\Program Files\nodejs\;C:\Program Files\MongoDB\Server\3.2\bin;C:\Program Files (x86)\Heroku\bin;C:\Program Files (x86)\git\cmd;C:\Program Files (x86)\Git\cmd;;C:\Program Files (x86)\Microsoft VS Code\bin;C:\Program Files;C:\Users\millerlj\AppData\Roaming\npm
+10 verbose lifecycle recipe-api@1.0.0~start: CWD: C:\Users\millerlj\Documents\GitHub\CSSE280-Recipe-Api
+11 silly lifecycle recipe-api@1.0.0~start: Args: [ '/d /s /c', 'node ./bin/www' ]
+12 silly lifecycle recipe-api@1.0.0~start: Returned: code: 1  signal: null
+13 info lifecycle recipe-api@1.0.0~start: Failed to exec start script
+14 verbose stack Error: recipe-api@1.0.0 start: `node ./bin/www`
+14 verbose stack Exit status 1
+14 verbose stack     at EventEmitter.<anonymous> (C:\Program Files\nodejs\node_modules\npm\lib\utils\lifecycle.js:255:16)
+14 verbose stack     at emitTwo (events.js:106:13)
+14 verbose stack     at EventEmitter.emit (events.js:191:7)
+14 verbose stack     at ChildProcess.<anonymous> (C:\Program Files\nodejs\node_modules\npm\lib\utils\spawn.js:40:14)
+14 verbose stack     at emitTwo (events.js:106:13)
+14 verbose stack     at ChildProcess.emit (events.js:191:7)
+14 verbose stack     at maybeClose (internal/child_process.js:877:16)
+14 verbose stack     at Process.ChildProcess._handle.onexit (internal/child_process.js:226:5)
+15 verbose pkgid recipe-api@1.0.0
+16 verbose cwd C:\Users\millerlj\Documents\GitHub\CSSE280-Recipe-Api
+17 error Windows_NT 6.1.7601
+18 error argv "C:\\Program Files\\nodejs\\node.exe" "C:\\Program Files\\nodejs\\node_modules\\npm\\bin\\npm-cli.js" "start"
+19 error node v6.9.0
+20 error npm  v3.10.8
+21 error code ELIFECYCLE
+22 error recipe-api@1.0.0 start: `node ./bin/www`
+22 error Exit status 1
+23 error Failed at the recipe-api@1.0.0 start script 'node ./bin/www'.
+23 error Make sure you have the latest version of node.js and npm installed.
+23 error If you do, this is most likely a problem with the recipe-api package,
+23 error not with npm itself.
+23 error Tell the author that this fails on your system:
+23 error     node ./bin/www
+23 error You can get information on how to open an issue for this project with:
+23 error     npm bugs recipe-api
+23 error Or if that isn't available, you can get their info via:
+23 error     npm owner ls recipe-api
+23 error There is likely additional logging output above.
+24 verbose exit [ 1, true ]
