+++ package.json
@@ -18,6 +18,7 @@
   },
   "homepage": "https://github.com/Jernik/CSSE280-Recipe-Api#readme",
   "dependencies": {
+    "bcryptjs":"^0.7.12",
     "body-parser": "^1.15.2",
     "cookie-parser": "^1.4.3",
     "debug": "^2.2.0",
@@ -26,6 +27,7 @@
     "method-override": "^2.3.6",
     "mongoose": "^4.6.5",
     "morgan": "^1.7.0",
+    "password-hash": "^1.2.2",
     "serve-favicon": "^2.3.0"
   }
 }
