+++ views/index.jade
@@ -1,0 +1,5 @@
+extends layout
+
+block content
+  h1= title
+  p Welcome to #{title}
